Update cypress tests to not depend on Netlify Functions
