+++ cypress.json
@@ -1,3 +1,3 @@
 {
-  "baseUrl": "http://localhost:8888"
+  "baseUrl": "http://localhost:3000"
 }
