+++ cypress/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-    "extends": [
-      "plugin:cypress/recommended"
-    ]
-  }
+  "extends": [
+    "plugin:cypress/recommended"
+  ]
+}
