+++ cypress/integration/home_spec.js
@@ -1,5 +1,5 @@
 describe('Home Page', () => {
-  before(() => {
+  beforeEach(() => {
     cy.intercept('.netlify/functions/scheduled-events', {
       fixture: 'get_scheduled_events_response_200.json',
     })
@@ -10,6 +10,13 @@ describe('Home Page', () => {
   })
 
   it('contains the next event date', () => {
+    const now = new Date('04/20/2021')
+    cy.clock(now.getTime())
+
+    cy.intercept('.netlify/functions/scheduled-events', {
+      fixture: 'get_scheduled_events_response_200.json',
+    })
+
     cy.visit('/')
 
     cy.contains('Friday, April 30, 2021')
