+++ src/pages/Home.tsx
@@ -9,6 +9,7 @@ import NoDataImage from '../images/no-data-illistration.jpg'
 import { Talk } from './Admin'
 import { ToastContainer, toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
+import { getNextScheduledEvent } from './getNextScheduledEvent'
 
 const talks: Talk[] = [
   // {
@@ -21,36 +22,6 @@ const talks: Talk[] = [
   // },
 ]
 
-const isToday = (date: Date) => {
-  return false
-  // const today = new Date()
-  // return (
-  //   date.getDate() === today.getDate() &&
-  //   date.getMonth() === today.getMonth() &&
-  //   date.getFullYear() === today.getFullYear()
-  // )
-}
-
-const getNextScheduledEvent = (dates: ScheduledEvent[]) => {
-  const now = new Date()
-  const MAX_DATE = 8640000000000000
-  let nextEvent: ScheduledEvent = { date: new Date(MAX_DATE).toString(), id: 0 }
-
-  dates.forEach((date) => {
-    const currentDate = new Date(date.date)
-    console.log(currentDate, currentDate >= now)
-
-    if (
-      (currentDate >= now || isToday(currentDate)) &&
-      currentDate < new Date(nextEvent.date)
-    ) {
-      nextEvent = date
-    }
-  })
-
-  return nextEvent
-}
-
 interface HomeProps {
   loggedIn: boolean
   user: netlifyIdentity.User | null
@@ -58,7 +29,7 @@ interface HomeProps {
   login: () => void
 }
 
-interface ScheduledEvent {
+export interface ScheduledEvent {
   date: string
   id: number
 }
@@ -77,7 +48,7 @@ type FormData = {
   time: string
 }
 
-const nullFormData = {
+const nullFormData: FormData = {
   date: '',
   name: '',
   title: '',
